Make the registry explicit with the obsidiandynamics/kafdrop:latest image (#91)
Don't just assume that `obsidiandynamics/kafdrop:latest` always points
to docker.io automatically. Be explicit.

Tested with:

    podman pull docker.io/obsidiandynamics/kafdrop:latest
    Trying to pull docker.io/obsidiandynamics/kafdrop:latest...
    Getting image source signatures
    Copying blob 86fbbbe1ee3f [=>------------------------------------] 10.3MiB / 172.0MiB
    ...

The reason this is needed is that when we install certain operators from
an IIB (Image Index Bundle), we also need to whitelist which registries
a machine is allowed to connect to and making this explicit makes this
work in a more obvious manner.