Refactor grafana
Cleaned up syncwaves and hooks, jobs. We do not manually wait for
resources to be up, we rely on argocd retries and eventual consistency.
Refactored the creation of prometheus datasource to a crd from a job.
Removed all the unneeded rbac permissions from the sa.
Added a consolelink for the grafana app for easier reachability since
this is one of the main components of the pattern.