Grafana plugin fix (#131)
* Change broken grafana dashboard plugin

The ryantxu-ajax-plugin is deprecated and no longer maintained. Recently it
has been removed from the latest version of grafana and gone from the
plugin catalog as well.
Same functionality can be achieved with dalvany-image-panel plugin.

* Remove unused image-server

Image-server was used only in grafana panels as a "proxy" between the
grafana dashboard and the s3 (ceph) storage.
The new plugin that is used in the dashboards does not work properly
with the image-server inbetween, so we directly fetching and using the
pictures from s3. The image-server application is obsolate and no longer
used.